Better name for function

+++ prime-router/build.gradle.kts
@@ -108,10 +108,12 @@ compileTestKotlin.kotlinOptions.jvmTarget = "11"
 tasks.clean {
     // Delete the old Maven build folder
     delete("target")
+    delete("../.environment/soap_service/soap/event/v1/*.event")
     // clean up all the old event files in the SOAP set up
     doLast {
-        val eventsDir = File("../.environment/soap_service/soap/event/v1/")
-        if (eventsDir.exists()) {
+        val eventsDir = File("../.environment/soap_service/soap/event/v1")
+        if (eventsDir.exists() && eventsDir.isDirectory && eventsDir.listFiles().isNotEmpty()) {
+            // Note FileUtils does not like when the folder is empty.
             FileUtils.listFiles(eventsDir, arrayOf("event"), true).forEach {
                 it.delete()
             }
